updates maps
Updated so it is possible to:
Make the maps full screen
Possible to search in the first map (if for instance know a port city from which successful departures are expected or similar...)
Switch on and off layers in the second map
Hover over the "successful trajectories" and get the information about their lenght and average speed (here we can use whatever attributes are in the geojson files)

Furthermore, another geojson file is added. The one with all trajectories in (originally) "old_trajectories". Displayed on the map but I think it is fine now that one can switch layers on and off.
Left to do for Tuesday: update with nicer geojson files – the latest result Victor has (after correction of lat/lon order). Think about if we should add an html table under which could have just mock-number currently. Just to show more output types then maps.