choose departure points
Two options:
- provide a list of lists with [lat, lon] in a python file.
-interactively selects departure points in map.
both options result in a geojson file with all points that is also displayed on the map where the user select the study area for the simulations.

left to consider:
is a python file what the user should provide, or rather for instance a csv-file? need to add how the  interactively selected points are saved for future use (in a python file with a list of list or rather a csv-file).